Tidy up beamline tests a bit more
