+++ tests/beamlines/test_SSX_expt.py
@@ -1,19 +1,10 @@
-import tempfile
-
 import numpy as np
 import pytest
 from numpy.testing import assert_array_equal
 
 from nexgen.beamlines.beamline_utils import PumpProbe
 from nexgen.beamlines.SSX_expt import run_extruder, run_fixed_target
-from nexgen.nxs_utils import Axis, TransformationType
-
-axes_list = [
-    Axis("omega", ".", TransformationType.ROTATION, (0, 0, -1)),
-    Axis("sam_z", "omega", TransformationType.TRANSLATION, (0, 0, 1)),
-    Axis("sam_y", "sam_z", TransformationType.TRANSLATION, (0, 1, 0)),
-    Axis("sam_x", "sam_y", TransformationType.TRANSLATION, (1, 0, 0)),
-]
+from tests.beamlines.conftest import axes_list
 
 test_pump = PumpProbe()
 test_pump.pump_status = True
@@ -56,27 +47,11 @@ def test_run_extruder_with_non_zero_omega():
     assert_array_equal(osc["omega"], np.repeat(90.0, 10))
 
 
-@pytest.fixture
-def dummy_chipmap_file():
-    lines = [
-        "01status    P3011       1\n",
-        "02status    P3021       0\n",
-        "03status    P3031       0\n",
-        "04status    P3041       0\n",
-    ]
-    test_map_file = tempfile.NamedTemporaryFile(
-        mode="w", suffix=".map", delete=False, encoding="utf-8"
-    )
-    with test_map_file as map:
-        map.writelines(lines)
-    yield test_map_file
-
-
-def test_run_fixed_target(dummy_chipmap_file):
+def test_run_fixed_target(dummy_chipmap_file_single_block):
     transl, info = run_fixed_target(
         axes_list,
         test_chip_dict,
-        dummy_chipmap_file.name,
+        dummy_chipmap_file_single_block.name,
         test_pump,
     )
     assert list(transl.keys()) == ["sam_y", "sam_x"]
@@ -85,66 +60,50 @@ def test_run_fixed_target(dummy_chipmap_file):
     assert info["n_exposures"] == 1
 
 
-def test_run_fixed_target_with_wrong_axis_in_list(dummy_chipmap_file):
+def test_run_fixed_target_with_wrong_axis_in_list(dummy_chipmap_file_single_block):
     with pytest.raises(ValueError):
         _ = run_fixed_target(
             axes_list,
             test_chip_dict,
-            dummy_chipmap_file.name,
+            dummy_chipmap_file_single_block.name,
             test_pump,
             scan_axes=["phi", "sam_x"],
         )
 
 
-def test_run_fixed_target_with_missing_scan_axis(dummy_chipmap_file):
+def test_run_fixed_target_with_missing_scan_axis(dummy_chipmap_file_single_block):
     with pytest.raises(IndexError):
         _ = run_fixed_target(
             axes_list,
             test_chip_dict,
-            dummy_chipmap_file.name,
+            dummy_chipmap_file_single_block.name,
             test_pump,
             scan_axes=["sam_z"],
         )
 
 
-def test_fixed_target_raises_error_if_no_chip_info(dummy_chipmap_file):
+def test_fixed_target_raises_error_if_no_chip_info(dummy_chipmap_file_single_block):
     with pytest.raises(ValueError):
         _ = run_fixed_target(
             axes_list,
             {},
-            dummy_chipmap_file.name,
+            dummy_chipmap_file_single_block.name,
             test_pump,
         )
 
 
-def test_fixed_target_for_multiple_exposures(dummy_chipmap_file):
+def test_fixed_target_for_multiple_exposures(dummy_chipmap_file_single_block):
     test_chip_dict["N_EXPOSURES"] = [0, "2"]
     transl, info = run_fixed_target(
         axes_list,
         test_chip_dict,
-        dummy_chipmap_file.name,
+        dummy_chipmap_file_single_block.name,
         test_pump,
     )
     assert list(transl.keys()) == ["sam_y", "sam_x"]
     assert len(transl["sam_x"]) == len(transl["sam_y"])
     assert len(transl["sam_y"]) == 800
     assert info["n_exposures"] == 2
-
-
-@pytest.fixture
-def dummy_chipmap_file_multi_block():
-    lines = [
-        "01status    P3011       1\n",
-        "02status    P3021       1\n",
-        "03status    P3031       1\n",
-        "04status    P3041       1\n",
-    ]
-    test_map_file = tempfile.NamedTemporaryFile(
-        mode="w", suffix=".map", delete=False, encoding="utf-8"
-    )
-    with test_map_file as map:
-        map.writelines(lines)
-    yield test_map_file
 
 
 def test_fixed_target_with_upwards_blocks(dummy_chipmap_file_multi_block):
